test(otlp): share span copy registry
